+++ CHANGELOG.md
@@ -18,6 +18,7 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
+- Fix kvstore instance class update error [GH-570]
 - Fix ess_scaling_group import bugs and improve ess schedule testcase [GH-565]
 - Fix alicloud rds related IncorrectStatus bug [GH-558]
 - Fix alicloud_fc_trigger's config diff bug [GH-556]